mir-system/requirements.txt
# Core audio processing
librosa==0.10.2
soundfile==0.12.1
pydub==0.25.1
noisereduce==3.0.2
pyloudnorm==0.1.1

# Deep learning
torch>=2.1.0
torchaudio>=2.1.0
timm==0.9.12

# Machine learning
scikit-learn==1.4.2
numpy>=1.26.0
scipy==1.13.0

# Similarity search
faiss-cpu==1.8.0

# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
python-multipart==0.0.9
httpx==0.27.0

# Database / ORM
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.20.0

# Task queue
celery[redis]==5.4.0
redis==5.0.4

# Object storage
boto3==1.34.100
minio==7.2.7

# Monitoring
prometheus-client==0.20.0
sentry-sdk[fastapi]==2.2.1

# Utilities
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
tenacity==8.3.0
structlog==24.1.0
rich==13.7.1
